% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{dms_saleOrder_del}
\alias{dms_saleOrder_del}
\title{删除DMS中的销售杨文意}
\usage{
dms_saleOrder_del(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FStartDate = "2022-08-01",
  FEndDate = "2022-08-31",
  FCompanyName = "苏州赛普生物科技有限公司"
)
}
\arguments{
\item{dms_token}{口令}

\item{FStartDate}{日期1}

\item{FEndDate}{日期2}

\item{FCompanyName}{公司名称}
}
\value{
返回值
}
\description{
删除DMS中的销售杨文意
}
\examples{
dms_saleOrder_del()
}
