% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOutReturnStock.R
\name{dms_saleOutStock_query}
\alias{dms_saleOutStock_query}
\title{数据中台订单查询}
\usage{
dms_saleOutStock_query(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FStartDate = "2022-08-01",
  FEndDate = "2022-08-31",
  FCompanyName = "苏州赛普生物科技有限公司"
)
}
\arguments{
\item{dms_token}{口令}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{FCompanyName}{公司名称}
}
\value{
返回值
}
\description{
数据中台订单查询
}
\examples{
dms_saleOrder_query()
}
