% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngrCost.R
\name{mngrCost_query}
\alias{mngrCost_query}
\title{查询数据}
\usage{
mngrCost_query(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FYear = 2022,
  FPeriod = 8
)
}
\arguments{
\item{dms_token}{口令}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
查询数据
}
\examples{
mngrCost_query()

}
