% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngrCost.R
\name{mngrCost_Server}
\alias{mngrCost_Server}
\title{查询管理成本}
\usage{
mngrCost_Server(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
查询管理成本
}
\examples{
mngrCost_Server()
}
