% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{organization_erp2dms}
\alias{organization_erp2dms}
\title{Title}
\usage{
organization_erp2dms(
  erp_token = "4D181CAB-4CE3-47A3-8F2B-8AB11BB6A227",
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{erp_token}{ERP口令}

\item{dms_token}{DMS口令}
}
\value{
返回值
}
\description{
Title
}
\examples{
organization_erp2dms()
}
