% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion.R
\name{promotion_query}
\alias{promotion_query}
\title{提成数据查询}
\usage{
promotion_query(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FSaleOrgName = "苏州赛普生物科技有限公司",
  FSaleManName = "陈龙",
  FStartDate = "2022-08-01",
  FEndDate = "2022-09-30"
)
}
\arguments{
\item{dms_token}{口令}

\item{FSaleOrgName}{销售组织}

\item{FSaleManName}{销售员}

\item{FStartDate}{开始结算日期}

\item{FEndDate}{as 结束结算日期}
}
\value{

}
\description{
提成数据查询
}
\examples{
promotion_query()
}
