% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion.R
\name{promotion_server}
\alias{promotion_server}
\title{提成查询}
\usage{
promotion_server(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
提成查询
}
\examples{
promotion_server()
}
