% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotionStat.R
\name{promotionEmp_Query}
\alias{promotionEmp_Query}
\title{提成数据查询}
\usage{
promotionEmp_Query(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FSaleOrgName = "苏州赛普生物科技有限公司",
  FYearPeriod = "202208"
)
}
\arguments{
\item{dms_token}{口令}

\item{FSaleOrgName}{销售组织}

\item{FSaleManName}{销售员}

\item{FStartDate}{开始结算日期}

\item{FEndDate}{as 结束结算日期}
}
\value{

}
\description{
提成数据查询
}
\examples{
promotionEmp_Query()
}
