% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleMan.R
\name{saleMan_query}
\alias{saleMan_query}
\title{查询销售员}
\usage{
saleMan_query(dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3")
}
\arguments{
\item{dms_token}{口令}
}
\value{
返回列表
}
\description{
查询销售员
}
\examples{
saleMan_query()
}
