% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{saleOrder_erp2dms}
\alias{saleOrder_erp2dms}
\title{Title}
\usage{
saleOrder_erp2dms(
  erp_token = "4D181CAB-4CE3-47A3-8F2B-8AB11BB6A227",
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FStartDate = "2022-08-01",
  FEndDate = "2022-08-31",
  FCompanyName = "苏州赛普生物科技有限公司"
)
}
\arguments{
\item{erp_token}{ERP口令}

\item{dms_token}{DMS口令}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{FCompanyName}{公司名称}
}
\value{
返回值
}
\description{
Title
}
\examples{
saleOrder_erp2dms()
}
