% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{server_saleOrder_click}
\alias{server_saleOrder_click}
\title{同步订单数据}
\usage{
server_saleOrder_click(input, output, session, erp_token, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{erp_token}{口令1}

\item{dms_token}{口令2}
}
\value{
返回值
}
\description{
同步订单数据
}
\examples{
server_saleOrder_click()
}
