% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOrder.R
\name{server_saleOrder_view}
\alias{server_saleOrder_view}
\title{同步订单数据}
\usage{
server_saleOrder_view(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
同步订单数据
}
\examples{
server_saleOrder_click()
}
