% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOutReturnStock_mfg.R
\name{server_saleOutStock_mfg_click}
\alias{server_saleOutStock_mfg_click}
\title{同步销售出库及销售退货数据,生产数据口径}
\usage{
server_saleOutStock_mfg_click(input, output, session, erp_token, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{erp_token}{口令1}

\item{dms_token}{口令2}
}
\value{
返回值
}
\description{
同步销售出库及销售退货数据,生产数据口径
}
\examples{
server_saleOutStock_mfg_click()
}
