% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOutReturnStock.R
\name{server_saleOutStock_view}
\alias{server_saleOutStock_view}
\title{同步出库数据}
\usage{
server_saleOutStock_view(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
同步出库数据
}
\examples{
server_saleOrder_click()
}
